# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SETTLE escrow platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sellers/register:
    post:
      operationId: registerSeller
      tags: [sellers]
      summary: Register a new seller account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerInput"
      responses:
        "201":
          description: Seller created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Seller"
        "400":
          description: Validation error
  /sellers/login:
    post:
      operationId: loginSeller
      tags: [sellers]
      summary: Login seller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellerLogin"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /deals/{id}/disputes:
    post:
      operationId: raiseDispute
      tags: [disputes]
      summary: Buyer raises a dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "201":
          description: Dispute raised, funds frozen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
  /deals/{id}/disputes/refund:
    post:
      operationId: issueRefund
      tags: [disputes]
      summary: Seller issues full refund to resolve dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Refund issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}/disputes/counter:
    post:
      operationId: submitCounterProof
      tags: [disputes]
      summary: Seller submits counter proof
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterProofInput"
      responses:
        "200":
          description: Counter proof submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dispute"
...
components:
  schemas:
...
    Seller:
      type: object
      required: [id, name, email, phone, walletAddress, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        walletAddress:
          type: string
        totalEarnings:
          type: number
        createdAt:
          type: string
    SellerInput:
      type: object
      required: [name, email, phone, password, walletAddress]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        walletAddress:
          type: string
    SellerLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, seller]
      properties:
        token:
          type: string
        seller:
          $ref: "#/components/schemas/Seller"
...
    Deal:
      type: object
      required: [id, code, shortUrl, itemName, description, price, deliveryWindowHours, status, sellerId, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        shortUrl:
          type: string
        itemName:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        feeAmount:
          type: number
        sellerPayout:
          type: number
        deliveryWindowHours:
          type: integer
        status:
          type: string
          enum: [created, locked, dispatched, delivered, settled, disputed]
        sellerId:
          type: integer
        buyerPhone:
          type: ["string", "null"]
        buyerName:
          type: ["string", "null"]
        dispatchedAt:
          type: ["string", "null"]
        deliveryDeadline:
          type: ["string", "null"]
        settledAt:
          type: ["string", "null"]
        dispute:
          $ref: "#/components/schemas/Dispute"
        createdAt:
          type: string
...
    Dispute:
      type: object
      required: [id, dealId, reason, status, createdAt]
      properties:
        id:
          type: integer
        dealId:
          type: integer
        reason:
          type: string
          enum: [item_never_arrived, wrong_damaged_item, incomplete_service]
        description:
          type: ["string", "null"]
        evidenceUrl:
          type: ["string", "null"]
        counterProofUrl:
          type: ["string", "null"]
        counterProofDescription:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, counter_submitted, resolved_refund, resolved_seller, escalated]
        resolution:
          type: ["string", "null"]
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    DisputeInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum: [item_never_arrived, wrong_damaged_item, incomplete_service]
        description:
          type: string
        evidenceUrl:
          type: string
    CounterProofInput:
      type: object
      required: [counterProofUrl]
      properties:
        counterProofUrl:
          type: string
        counterProofDescription:
          type: string
...