# lib/api-zod/src/generated/types/counterProofInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SETTLE escrow platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CounterProofInput {
  counterProofUrl: string;
  counterProofDescription?: string;
}