# lib/api-zod/src/generated/types/seller.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SETTLE escrow platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Seller {
  id: number;
  name: string;
  email: string;
  phone: string;
  walletAddress: string;
  totalEarnings?: number;
  createdAt: string;
}